Refactoring and adding tests
