+++ jvm/src/main/scala/com/cibo/evilplot/geometry/Graphics2DSupport.scala
@@ -262,7 +262,7 @@ final case class Graphics2DRenderContext(graphics: Graphics2D)
     }
   }
 
-  // to implement, reference: InteractiveCanvasContext.scala and CanvasInteractionMask.scala
+  // to implement, reference: InteractionMask.scala and InteractiveCanvasContext.scala
   override def draw(interaction: Interaction): Unit = interaction.r.draw(this)
 }
